Added support for a Definition Import column
If a sheet contains a "Definition URI" column, the "derived from"
property of the SynBioHub entry will be synchronized with the
value in "Definition URI" column.  If an empty "Definition URI"
column is added to an existing sheet it will cause all existing
"derived from" properties of the sheet entries in SynBioHub to
be cleared.  The "Definition Import" column resolves this issue.
If "Definition Import" column is added to a sheet that does not
contain a "Defitition URI" column, it will be loaded with the
"derived from" entries in SynBioHub.  Then it can be safely
renamed to be a "Definition URI" column.