Re-wrote tab-copy method
The tab copy method previously did not behave properly if the
destination spreadsheet already had a tab with the same name
as the tab being copied.

Reorded operations a bit to rename old tab in destination
spreadsheet prior to copying, ensuring at least one recent version
of the tab is saved in case the copy operation fails.